# httpapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Pit Officer REST API
  version: "1"
  description: >
    Control-plane REST API for the Pit Officer service. The public surface is
    mounted at `/api/v1`; a mirror at `/app/api/v1` is reserved for the operator
    dashboard and carries the same routes except the route manifest, which is
    published only on `/api/v1`. Control endpoints answer `application/json`;
    endpoints that hand back a file answer with that file's own media type.
    A malformed request body, or one that violates the schema or a parameter
    constraint, answers RFC 9457 `application/problem+json` - `400` for broken
    syntax, `422` for a constraint violation. Every other error answers the
    envelope `{"error":{"code":"...","message":"..."}}`.
    Money and size values
    are exact decimal strings - never IEEE 754 floats.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Public REST API
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      summary: List all assets
      tags: [assets]
      parameters:
        - $ref: "#/components/parameters/CodeFilter"
        - $ref: "#/components/parameters/CodeMatch"
        - name: class
          in: query
          description: Asset class code search text.
          schema:
            type: string
        - name: classMatch
          in: query
          description: Anchoring mode applied to the class filter.
          schema:
            type: string
            enum: [contains, starts_with, ends_with, exact]
            default: contains
        - name: sort
          in: query
          schema:
            type: string
            enum: [code, title, assetClass]
          description: Asset list sort key. Unknown keys are rejected.
        - $ref: "#/components/parameters/SortOrder"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: Asset list.
          content:
            application/json:
              schema:
                type: object
                required: [assets, total]
                properties:
                  assets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Asset"
                  total:
                    type: integer
                    description: Total assets matching the filters before paging.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createAsset
      summary: Create an asset
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  type: string
                  description: Operator-chosen asset public code.
                title:
                  type: string
                  maxLength: 256
                  description: Mutable display title for the asset.
                assetClass:
                  type: string
                  maxLength: 256
                  description: Optional operator classification for the asset.
      responses:
        "201":
          description: Asset created.
          content:
            application/json:
              schema:
                type: object
                required: [asset]
                properties:
                  asset:
                    $ref: "#/components/schemas/Asset"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /assets/{code}:
    parameters:
      - name: code
        in: path
        required: true
        description: URL-encoded asset public code.
        schema:
          type: string
    put:
      operationId: updateAsset
      summary: Update an asset's code, title and classification
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  type: string
                  description: >
                    Replacement asset public code. Renaming updates the stored
                    asset and records the change in the audit log.
                title:
                  type: string
                  maxLength: 256
                  description: Mutable display title for the asset.
                assetClass:
                  type: string
                  maxLength: 256
                  description: Optional operator classification for the asset.
      responses:
        "200":
          description: Asset after update.
          content:
            application/json:
              schema:
                type: object
                required: [asset]
                properties:
                  asset:
                    $ref: "#/components/schemas/Asset"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "503":
          $ref: "#/components/responses/EngineRestartingError"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteAsset
      summary: Delete an asset
      tags: [assets]
      parameters:
        - $ref: "#/components/parameters/ForceCascade"
      responses:
        "204":
          description: Asset deleted.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/HasDependentsError"
        "500":
          $ref: "#/components/responses/InternalError"
  /asset-classes:
    get:
      operationId: listAssetClasses
      summary: List all asset classes
      tags: [assets]
      parameters:
        - $ref: "#/components/parameters/CodeFilter"
        - $ref: "#/components/parameters/CodeMatch"
        - name: notes
          in: query
          schema:
            type: string
          description: Search asset-class notes.
        - $ref: "#/components/parameters/NotesMatch"
        - name: sort
          in: query
          schema:
            type: string
            enum: [code, title, assetCount]
          description: Asset-class list sort key. Unknown keys are rejected.
        - $ref: "#/components/parameters/SortOrder"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: Asset-class list.
          content:
            application/json:
              schema:
                type: object
                required: [assetClasses, total]
                properties:
                  assetClasses:
                    type: array
                    items:
                      $ref: "#/components/schemas/AssetClass"
                  total:
                    type: integer
                    description: Total asset classes matching the filters before paging.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createAssetClass
      summary: Create an asset class
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  type: string
                  description: Asset-class public code.
                title:
                  type: string
                  maxLength: 256
                  description: Mutable display title for the class.
                notes:
                  type: string
                  maxLength: 4096
                  description: Free-form operator notes stored on the class.
      responses:
        "201":
          description: Asset class created.
          content:
            application/json:
              schema:
                type: object
                required: [assetClass]
                properties:
                  assetClass:
                    $ref: "#/components/schemas/AssetClass"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /asset-classes/{code}:
    parameters:
      - name: code
        in: path
        required: true
        description: URL-encoded asset-class public code.
        schema:
          type: string
    put:
      operationId: updateAssetClass
      summary: Update an asset class's code, title and notes
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  type: string
                  description: >
                    Replacement asset-class public code. A rename cascades the
                    asset.assetClass text link to referencing assets.
                title:
                  type: string
                  maxLength: 256
                  description: Replacement display title.
                notes:
                  type: string
                  maxLength: 4096
                  description: Replacement operator notes.
      responses:
        "200":
          description: Asset class after update.
          content:
            application/json:
              schema:
                type: object
                required: [assetClass]
                properties:
                  assetClass:
                    $ref: "#/components/schemas/AssetClass"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteAssetClass
      summary: Delete an asset class
      tags: [assets]
      parameters:
        - $ref: "#/components/parameters/ForceCascade"
      responses:
        "204":
          description: Asset class deleted.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/HasDependentsError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  parameters:
...
    CodeFilter:
      name: code
      in: query
      description: >
        Code search text. A plain value is matched as a substring. `*` is a
        friendly wildcard whose fragments must appear in order.
      schema:
        type: string
    CodeMatch:
      name: codeMatch
      in: query
      description: Anchoring mode applied to the code filter.
      schema:
        type: string
        enum: [contains, starts_with, ends_with, exact]
        default: contains
    NotesMatch:
      name: notesMatch
      in: query
      description: Anchoring mode applied to the notes filter.
      schema:
        type: string
        enum: [contains, starts_with, ends_with, exact]
        default: contains
...
    SortOrder:
      name: order
      in: query
      description: Sort direction.
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    ForceCascade:
      name: force
      in: query
      description: Allow application-approved cascade deletion of dependent rows.
      schema:
        type: boolean
        default: false
...
    LimitParam:
      name: limit
      in: query
      description: Maximum rows to return (default 100, capped at 1000).
      schema:
        type: integer
        default: 100
        maximum: 1000
    OffsetParam:
      name: offset
      in: query
      description: Zero-based row offset for server-side pagination.
      schema:
        type: integer
        default: 0
        minimum: 0
    PageParam:
      name: page
      in: query
      description: >
        One-based page number for server-side pagination. When present, the
        backend computes offset as `(page - 1) * limit`.
      schema:
        type: integer
        minimum: 1

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    MalformedRequestError:
      description: Malformed HTTP request or JSON syntax.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationError:
      description: Syntactically valid request failed schema or parameter validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ConflictError:
      description: >-
        Resource conflict, including an existing code, protected runtime state,
        or a configured dependency that prevents the mutation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    HasDependentsError:
      description: Delete would remove dependent rows without force.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HasDependentsError"
...
    EngineRestartingError:
      description: Engine is being rebuilt; mutating requests are rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, errors]
      properties:
        type:
          type: string
          description: >
            RFC 9457 problem type URI. Officer uses `about:blank` for generic
            HTTP errors.
          enum: ["about:blank"]
        title:
          type: string
          description: Canonical HTTP status title.
        status:
          type: integer
          enum: [400, 422]
        detail:
          type: string
          description: Human-readable explanation for this occurrence.
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ProblemError"
    ProblemError:
      type: object
      additionalProperties: false
      required: [code, pointer]
      properties:
        code:
          type: string
          description: Stable machine-readable error category.
        constraint:
          type: string
          description: Violated request constraint, when known.
        pointer:
          type: string
          description: JSON Pointer to the invalid member; empty means the request root.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: >
                Machine-readable code: `validation`, `not_found`,
                `account_missing`, `conflict`, `currency_change_blocked`, `has_dependents`,
                `terminal_order`, `execution_report_required`, `precondition`,
                `too_large`, `forbidden`, `reserved_group`,
                `engine_restarting`, `signing`, `upstream`, `not_implemented`,
                or `internal`.
              example: not_found
            message:
              type: string
              example: account not found
...
    DependentCount:
      type: object
      required: [kind, count]
      properties:
        kind:
          type: string
          description: Dependent row category.
        count:
          type: integer
          minimum: 0
          description: Number of dependent rows in this category.
    HasDependentsError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, dependents]
          properties:
            code:
              type: string
              enum: [has_dependents]
            message:
              type: string
              example: has dependents
            dependents:
              type: array
              items:
                $ref: "#/components/schemas/DependentCount"
...
    Asset:
      type: object
      required: [code, title, assetClass]
      properties:
        code:
          type: string
          description: >
            Operator-chosen public code, renameable through asset update.
        title:
          type: string
          description: Mutable display title.
        assetClass:
          type: string
          description: Optional operator classification for the asset.

    # AssetClass
    AssetClass:
      type: object
      required: [code, title, notes, assetCount]
      properties:
        code:
          type: string
          description: Asset-class public code (its handle).
        title:
          type: string
          description: Mutable display title.
        notes:
          type: string
          description: Free-form operator notes.
        assetCount:
          type: integer
          minimum: 0
          description: Number of assets referencing this class.

    # Account
...